requirements.txt
tkinter
pycryptodome
socket
bcrypt